refactor: 모듈형 디렉토리로 변경
