+++ __tests__/domain/shared/domain/Lotto.test.js
@@ -1,4 +1,5 @@
-import Lotto from "../../../src/domain/models/Lotto.js";
+import Lotto from "../../../../src/lotto/shared/domain/Lotto";
+import WinningNumber from "../../../../src/lotto/shared/domain/WinningNumber";
 
 describe("Lotto 생성 성공테스트", () => {
   test("로또번호 성공 테스트", () => {
