+++ src/lotto/shared/domain/WinningNumber.js
